requirements.txt
request
black
flake8
mypy